Добавил Header и вынес загрузку в арр
